requirements.txt
streamlit
streamlit-pdf-viewer
python-dotenv
pandas
tqdm
matplotlib
# requests
pyocr
PyPDF2